some refactoring to shut code climate up
